% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlexswesent.R
\docType{package}
\name{rlexswesent-package}
\alias{rlexswesent}
\alias{rlexswesent-package}
\title{rlexswesent R package}
\description{
An convinience R package that contain the Swedish sentiment lexicon from Språkbanken, 
together with functionality to use with tidytext formats and dplyr.
}

