% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{swesent}
\alias{swesent}
\title{Sentiment lexicon for Swedish}
\format{A \code{tbl_df} with one row per lemgram.}
\usage{
swesent
}
\description{
Sentiment lexicon for Swedish
}
\references{
\url{http://www.ep.liu.se/ecp/126/006/ecp16126006.pdf}
}
\keyword{datasets}

