% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_written_form.R
\name{tidy_written_form}
\alias{tidy_written_form}
\title{Return the different written forms as a \code{data_frame}}
\usage{
tidy_written_form(swesent)
}
\arguments{
\item{swesent}{The \code{swesent}}
}
\description{
Return the different written forms as a \code{data_frame}
}
